simple multiple files

+++ src/element/file.tsx
@@ -3,8 +3,9 @@ import {
   Trash as TrashIcon,
 } from 'react-bootstrap-icons';
 import React from 'react';
-import { Button, Col } from 'react-bootstrap';
+import { Button, Col, Form } from 'react-bootstrap';
 import { FileProps, FileState } from './index';
+import ConfirmButton from 'components/button/ConfirmButton';
 
 export default class FileComponent extends React.Component<
   FileProps,
@@ -12,6 +13,9 @@ export default class FileComponent extends React.Component<
 > {
   constructor(props: FileProps) {
     super(props);
+    this.state = {
+      showDelete: false,
+    };
   }
 
   render() {
@@ -24,20 +28,64 @@ export default class FileComponent extends React.Component<
         xxl={2}
         className="px-2 mt-3 position-relative file"
         onClick={() => {
-          this.props.onSelect(this.props.file);
+          this.props.onSelect(
+            Array.isArray(this.props.selectedFile)
+              ? this.props.selectedFile.filter(
+                  (e) => e.id === this.props.file.id
+                ).length > 0
+                ? null
+                : this.props.file
+              : // If not multiple
+              this.props.file.id === this.props.selected?.id
+              ? null
+              : this.props.file
+          );
+        }}
+        onMouseEnter={() => {
+          this.setState({
+            showDelete: true,
+          });
+        }}
+        onMouseLeave={() => {
+          this.setState({
+            showDelete: false,
+          });
         }}
       >
-        <Button
-          variant="danger"
-          className="position-absolute remove-file"
-          style={{ top: '4px', right: '12px', lineHeight: 1, padding: '2px' }}
-          onClick={(e) => {
-            e.stopPropagation();
-            if (this.props.onDelete) this.props.onDelete(this.props.file);
-          }}
-        >
-          <TrashIcon />
-        </Button>
+        {this.state.showDelete && (
+          <ConfirmButton
+            title="Confirm?"
+            content="Delete this file?"
+            confirm={{
+              textConfirm: 'Delete',
+              variantConfirm: 'danger',
+              onConfirm: (e) => {
+                if (this.props.onDelete) this.props.onDelete(this.props.file);
+              },
+            }}
+            className="position-absolute remove-file"
+            style={{ top: '4px', right: '12px' }}
+          >
+            <Button
+              variant="danger"
+              className="w-100"
+              style={{ lineHeight: 1, padding: '2px' }}
+            >
+              <TrashIcon />
+            </Button>
+          </ConfirmButton>
+        )}
+        {(Array.isArray(this.props.selectedFile)
+          ? this.props.selectedFile.filter((e) => e.id === this.props.file.id)
+              .length > 0
+          : this.props.file.id === this.props.selected?.id) && (
+          <Form.Check
+            type="checkbox"
+            defaultChecked
+            className="position-absolute"
+            style={{ top: '4px', left: '16px' }}
+          />
+        )}
         <div
           style={{
             width: '100%',
